Refactor[#55]: Refactor playlist_screen & folders
