+++ catalogo_filmes/lib/components/edit_playlist.dart
@@ -1,13 +1,8 @@
-import 'dart:math';
-
-import 'package:catalogo_filmes/models/playlist.dart';
-import 'package:catalogo_filmes/providers/playlists_provider.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
+
+import '../models/playlist.dart';
+import '../providers/playlists_provider.dart';
 
 class EditPlaylist extends StatefulWidget {
   Playlist playlist;
@@ -70,6 +65,7 @@ class _EditPlaylistState extends State<EditPlaylist> {
           child: Form(
               key: _formKey,
               child: Column(
+                mainAxisSize: MainAxisSize.min,
                 children: [
                   TextFormField(
                     initialValue: _formData['name'],
