+++ catalogo_filmes/lib/components/movie_items/favorite_movie_item.dart
@@ -1,13 +1,8 @@
-import 'dart:math';
-
-import 'package:catalogo_filmes/providers/favorites_provider.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../models/movie.dart';
+import '/providers/favorites_provider.dart';
+import '../../models/movie.dart';
 
 class CardMovieItem extends StatefulWidget {
   Movie movie;
@@ -51,6 +46,7 @@ class _CardMovieItemState extends State<CardMovieItem> {
                 right: 3,
               ),
               Positioned(
+                top: 0,
                 child: Container(
                   color: Colors.black45,
                   padding: const EdgeInsets.only(left: 10, right: 10, top: 15),
@@ -78,66 +74,10 @@ class _CardMovieItemState extends State<CardMovieItem> {
                     ],
                   ),
                 ),
-                top: 0,
               )
             ],
           ),
         ),
-        // Expanded(
-        //   flex: 1,
-        //   child: Container(
-        //     padding: const EdgeInsets.only(left: 10),
-        //     color: Theme.of(context).colorScheme.primary,
-        //     height: MediaQuery.of(context).size.height * 0.4,
-        //     width: MediaQuery.of(context).size.width,
-        //     child: ListView(children: [
-        //       Column(
-        //         crossAxisAlignment: CrossAxisAlignment.start,
-        //         children: [
-        //           Text(
-        //             'Título Completo: ${widget.movie.title}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //           const SizedBox(
-        //             height: 20,
-        //           ),
-        //           Text(
-        //             'Ano: ${widget.movie.year}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //           const SizedBox(
-        //             height: 20,
-        //           ),
-        //           Text(
-        //             'Diretores: ${widget.movie.directors}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //           const SizedBox(
-        //             height: 20,
-        //           ),
-        //           Text(
-        //             'Elenco: ${widget.movie.crew}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //           const SizedBox(
-        //             height: 20,
-        //           ),
-        //           Text(
-        //             'Nota IMDb: ${widget.movie.rate}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //           const SizedBox(
-        //             height: 20,
-        //           ),
-        //           Text(
-        //             'Sinopse: ${widget.movie.plot}',
-        //             style: Theme.of(context).textTheme.headline2,
-        //           ),
-        //         ],
-        //       ),
-        //     ]),
-        //   ),
-        // ),
       ],
     );
   }
