+++ catalogo_filmes/lib/components/movie_items/movie_menu.dart
@@ -1,15 +1,11 @@
-import 'package:catalogo_filmes/components/playlist_menu.dart';
-import 'package:catalogo_filmes/components/rating_form.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../models/movie.dart';
-import '../providers/favorites_provider.dart';
-import '../providers/playlists_provider.dart';
-import 'new_playlist.dart';
+import '../../models/movie.dart';
+import '../../providers/favorites_provider.dart';
+import '../../providers/playlists_provider.dart';
+import '../new_playlist.dart';
+import '../playlist_menu.dart';
 
 enum Options { create, addTo, rating }
